requirements.txt
streamlit==1.32.0
librosa==0.10.1
numpy==1.26.4
scipy==1.12.0
soundfile==0.12.1
scikit-learn==1.4.2
python-dotenv==1.1.0
requests==2.31.0
transformers==4.38.2
torch==2.2.1

